+++ actionpack/CHANGELOG.md
@@ -1,3 +1,9 @@
+*   Replace "blocked host" error page with generic 403 Forbidden in production environment.
+
+    Fixes #42813.
+
+    *Ajax Manohar*
+
 *   Deprecate `poltergeist` and `webkit` (capybara-webkit) driver registration for system testing (they will be removed in Rails 7.1). Add `cuprite` instead.
 
     [Poltergeist](https://github.com/teampoltergeist/poltergeist) and [capybara-webkit](https://github.com/thoughtbot/capybara-webkit) are already not maintained. These usage in Rails are removed for avoiding confusing users.